chore: renamed headers to hpp style
