+++ lib/status.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
 
-typedef enum
+enum Status_t
 {
     kStatusSuccess = 0x00,
     kStatusOutOfMemory,
     kStatusFailure,
     kStatusInvalidParam
-} Status_t;
+} ;
